+++ frontend/src/components/EmployeeDetails.js
@@ -1,10 +1,20 @@
-import { useEmployeesContext } from "../hooks/useEmployeeContext"
+import { useEmployeesContext } from '../hooks/useEmployeeContext'
+import { useAuthContext } from '../hooks/useAuthContext'
 
 const EmployeeDetails = ({employee}) => {
-    const {dispatch} = useEmployeesContext()
+    const { dispatch } = useEmployeesContext()
+    const { user } = useAuthContext()
+
     const handleClickDelete = async () => {
+        if (!user) {
+            return
+        }
+
         const response = await fetch('/api/employees/' + employee._id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json()
